requirements.txt
pandas
numpy
yfinance
psycopg2-binary
